+++ bedoner/pipelines/udify.py
@@ -8,7 +8,7 @@ from bedoner.vendor.udify.models.udify_model import OUTPUTS as UdifyOUTPUTS
 from spacy.tokens import Doc, Token
 
 from .allennlp_base import AllennlpPipe
-from .utils import set_heads
+from .utils import flatten_docs_to_sents, set_heads
 
 spacy.language.ENABLE_PIPELINE_ANALYSIS = True
 
@@ -20,25 +20,25 @@ import_submodules("bedoner.vendor.udify")
 )
 class Udify(AllennlpPipe):
     def set_annotations(self, docs: Iterable[Doc], outputs: Dict):
-        for doc, output in zip(docs, outputs):
+        for sent, output in zip(flatten_docs_to_sents(docs), outputs):
             deps = output[UdifyOUTPUTS.predicted_dependencies]
             heads = output[UdifyOUTPUTS.predicted_heads]
             uposes = output[UdifyOUTPUTS.upos]
             lemmas = output[UdifyOUTPUTS.lemmas]
             words = output[UdifyOUTPUTS.words]
-            _doc_tokens = [token.text for token in doc]
-            if not words == [token.text for token in doc]:
+            _doc_tokens = [token.text for token in sent]
+            if not words == [token.text for token in sent]:
                 raise ValueError(
                     "Internal error has occured."
-                    f"Input text: {doc.text}\n"
+                    f"Input text: {sent.text}\n"
                     f"Input tokens: {_doc_tokens}\n"
                     f"Model words: {words}"
                 )
 
-            for token, dep, upos, lemma in zip(doc, deps, uposes, lemmas):
+            for token, dep, upos, lemma in zip(sent, deps, uposes, lemmas):
                 token: Token = token
                 token.dep_ = dep
                 token.lemma_ = lemma
                 token.pos_ = upos
-            doc = set_heads(doc, heads)
-            doc.is_parsed = True
+            sent = set_heads(sent, heads)
+            sent.doc.is_parsed = True